+++ contracts/liquidity_book/tests/src/multitests/lb_token.rs
@@ -130,7 +130,7 @@ pub fn test_simple_mint() -> Result<(), anyhow::Error> {
         liquidity_parameters,
     )?;
 
-    let total_bins = get_total_bins(nb_bins_x, nb_bins_y) as u32;
+    let total_bins = get_total_bins(nb_bins_x as u32, nb_bins_y as u32) as u32;
 
     for i in 0..total_bins {
         let id = get_id(ACTIVE_ID, i, nb_bins_y);
@@ -205,7 +205,7 @@ pub fn test_mint_twice() -> Result<(), anyhow::Error> {
         nb_bins_x,
         nb_bins_y,
     )?;
-    let total_bins = get_total_bins(nb_bins_x, nb_bins_y) as u32;
+    let total_bins = get_total_bins(nb_bins_x as u32, nb_bins_y as u32) as u32;
 
     lb_pair::add_liquidity(
         &mut app,
@@ -321,7 +321,7 @@ pub fn test_mint_with_different_bins() -> Result<(), anyhow::Error> {
         liquidity_parameters,
     )?;
 
-    let total_bins = get_total_bins(nb_bins_x, nb_bins_y) as u32;
+    let total_bins = get_total_bins(nb_bins_x as u32, nb_bins_y as u32) as u32;
     let mut balances = vec![Uint256::zero(); total_bins as usize];
 
     for i in 0..total_bins {
@@ -400,7 +400,7 @@ pub fn test_mint_with_different_bins() -> Result<(), anyhow::Error> {
         }
     }
 
-    let total_bins = get_total_bins(nb_bins_x, nb_bins_y) as u32;
+    let total_bins = get_total_bins(nb_bins_x as u32, nb_bins_y as u32) as u32;
 
     for i in 0..total_bins {
         let id = get_id(ACTIVE_ID, i, nb_bins_y);
@@ -479,7 +479,7 @@ pub fn test_simple_burn() -> Result<(), anyhow::Error> {
         liquidity_parameters,
     )?;
 
-    let total_bins = get_total_bins(nb_bins_x, nb_bins_y) as u32;
+    let total_bins = get_total_bins(nb_bins_x as u32, nb_bins_y as u32) as u32;
     let mut balances = vec![Uint256::zero(); total_bins as usize];
     let mut ids = vec![0u32; total_bins as usize];
 
@@ -519,7 +519,7 @@ pub fn test_simple_burn() -> Result<(), anyhow::Error> {
     assert_eq!(reserves_x, 0u128);
     assert_eq!(reserves_y, 0u128);
 
-    let total_bins = get_total_bins(nb_bins_x, nb_bins_y) as u32;
+    let total_bins = get_total_bins(nb_bins_x as u32, nb_bins_y as u32) as u32;
 
     for i in 0..total_bins {
         let id = get_id(ACTIVE_ID, i, nb_bins_y);
@@ -535,7 +535,7 @@ pub fn test_simple_burn() -> Result<(), anyhow::Error> {
         assert_eq!(balance, Uint256::zero());
     }
 
-    let total_bins = get_total_bins(nb_bins_x, nb_bins_y) as u32;
+    let total_bins = get_total_bins(nb_bins_x as u32, nb_bins_y as u32) as u32;
 
     for i in 0..total_bins {
         let id = get_id(ACTIVE_ID, i, nb_bins_y);
@@ -613,7 +613,7 @@ pub fn test_burn_half_twice() -> Result<(), anyhow::Error> {
         liquidity_parameters,
     )?;
 
-    let total_bins = get_total_bins(nb_bins_x, nb_bins_y) as u32;
+    let total_bins = get_total_bins(nb_bins_x as u32, nb_bins_y as u32) as u32;
     let mut balances = vec![Uint256::zero(); total_bins as usize];
     let mut half_balances = vec![Uint256::zero(); total_bins as usize];
     let mut ids = vec![0u32; total_bins as usize];
